platform/snp: Add abstractions for FW operations
Add new abstractions to SvsmPlatform:

- prepare_fw() performs preparations for launching guest boot firmware
- launch_fw() launches guest boot firmware

Move FW-related functions in svsm.rs to snp_fw.rs as they're
SEV-specific. Make sure snp_fw doesn't export anything unnecessary to
the outside world.
