+++ kernel/src/igvm_params.rs
@@ -11,8 +11,7 @@ use crate::address::{PhysAddr, VirtAddr};
 use crate::cpu::efer::EFERFlags;
 use crate::error::SvsmError;
 use crate::mm::{GuestPtr, PerCPUPageMappingGuard, PAGE_SIZE};
-use crate::platform::snp_fw::SevFWMetaData;
-use crate::platform::{PageStateChangeOp, PageValidateOp, SVSM_PLATFORM};
+use crate::platform::{PageStateChangeOp, PageValidateOp, SevFWMetaData, SVSM_PLATFORM};
 use crate::types::PageSize;
 use crate::utils::MemoryRegion;
 use alloc::vec::Vec;
