+++ kernel/src/platform/mod.rs
@@ -7,8 +7,10 @@
 pub mod guest_cpu;
 pub mod native;
 pub mod snp;
-pub mod snp_fw;
 pub mod tdp;
+
+mod snp_fw;
+pub use snp_fw::{parse_fw_meta_data, SevFWMetaData};
 
 use native::NativePlatform;
 use snp::SnpPlatform;
@@ -17,6 +19,7 @@ use tdp::TdpPlatform;
 use core::ops::{Deref, DerefMut};
 
 use crate::address::{PhysAddr, VirtAddr};
+use crate::config::SvsmConfig;
 use crate::cpu::cpuid::CpuidResult;
 use crate::cpu::percpu::PerCpu;
 use crate::error::SvsmError;
@@ -75,6 +78,20 @@ pub trait SvsmPlatform {
     /// Performs initialiation of the environment specfic to the SVSM kernel
     /// (for services not used by stage2).
     fn env_setup_svsm(&self) -> Result<(), SvsmError>;
+
+    /// Performs the necessary preparations for launching guest boot firmware.
+    fn prepare_fw(
+        &self,
+        _config: &SvsmConfig<'_>,
+        _kernel_region: MemoryRegion<PhysAddr>,
+    ) -> Result<(), SvsmError> {
+        Ok(())
+    }
+
+    /// Launches guest boot firmware.
+    fn launch_fw(&self, _config: &SvsmConfig<'_>) -> Result<(), SvsmError> {
+        Ok(())
+    }
 
     /// Completes initialization of a per-CPU object during construction.
     fn setup_percpu(&self, cpu: &PerCpu) -> Result<(), SvsmError>;
